refactor(getUnhandledProps): move to propUtils
