+++ src/collections/Form/FormFields.js
@@ -1,7 +1,7 @@
 import _ from 'lodash'
 import React, { Children, Component, PropTypes } from 'react'
 import classNames from 'classnames'
-import getUnhandledProps from '../../utils/getUnhandledProps'
+import { getUnhandledProps } from '../../utils/propUtils'
 import numberToWord from '../../utils/numberToWord'
 import META from '../../utils/Meta.js'
