+++ src/utils/defaultProps.js
@@ -1,0 +1,15 @@
+import { PropTypes } from 'react'
+
+const defaultProps = {
+  /**
+   * Child components.
+   */
+  children: PropTypes.node,
+
+  /**
+   * Additional component classes appended to the root component.
+   */
+  className: PropTypes.string,
+}
+
+export default defaultProps
