+++ src/utils/propUtils.js
@@ -21,6 +21,19 @@ export const getComponentProps = (props, pluginPropTypes) => {
   return _.omit(props, _.keys(pluginPropTypes))
 }
 
+/**
+ * Returns an object consisting of props not defined in propTypes unless defined in defaultProps.
+ * @param {*} instance The `this` keyword in a React Component class.
+ * @returns {{}} A shallow copy of the prop object
+ */
+export const getUnhandledProps = (instance) => {
+  return _.omit(instance.props, (val, key) => {
+    const inPropTypes = _.has(instance.constructor.propTypes, key)
+    const inDefaultProps = _.has(instance.constructor.defaultProps, key)
+    return inPropTypes && !inDefaultProps
+  })
+}
+
 export const customPropTypes = {
   /*
    * Ensures children are of a set of types
