+++ src/views/Statistic/StatisticValue.js
@@ -1,6 +1,6 @@
 import React, { Component, PropTypes } from 'react'
 import classNames from 'classnames'
-import getUnhandledProps from '../../utils/getUnhandledProps'
+import { getUnhandledProps } from '../../utils/propUtils'
 import META from '../../utils/Meta'
 
 export default class StatisticValue extends Component {
